requirements.txt
Click==7.0
colorgram.py==1.2.0
Flask==1.1.0
itsdangerous==1.1.0
Jinja2==2.10.1
MarkupSafe==1.1.1
Pillow==6.1.0
Werkzeug==0.15.4
